feat(SnsShare): unify of SnsLink with SnsShare
